requirements.txt
langchain
langchain-openai
python-dotenv
langchain-google-genai
streamlit
langchain-ollama
langchain-groq
langchain-community 
langgraph
pymupdf
python-docx
pandas
python-pptx
openpyxl
youtube-transcript-api
